requirements.txt
rembg==2.0.68
flask==3.1.2
pillow==12.0.0
requests==2.32.5
